workers,trace_events: set thread name for workers
Set the thread name for workers in trace events. Also,
use uint64_t for thread_id_ because there's really no
reason for those to be doubles

PR-URL: https://github.com/nodejs/node/pull/21246

+++ src/node_worker.h
@@ -62,7 +62,7 @@ class Worker : public AsyncWrap {
 
   bool thread_joined_ = true;
   int exit_code_ = 0;
-  double thread_id_ = -1;
+  uint64_t thread_id_ = -1;
 
   std::unique_ptr<MessagePortData> child_port_data_;
